+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Apply code style related fixes
 - Only validate language in FrontController when entity is translatable
 - Allow early rendering in controllers
+- Allow adding attachments to ViewBuilder
 
 ### Removed
 - Remove dependency on the node module